tests/wpt/web-platform-tests/clipboard-apis/events/copy-event-manual.html

<!DOCTYPE html>
<title>The copy event</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#clipboard-event-copy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<p>Select and copy any part of this text to continue.
<script>
setup({explicit_timeout: true});
async_test(t => {
  document.oncopy = t.step_func_done(event => {
    // Nothing can be asserted about the event target until
    // https://github.com/w3c/clipboard-apis/issues/70 is resolved.
    // assert_equals(event.target, document.body, 'event.target');
    assert_true(event.isTrusted, 'event.isTrusted');
    assert_true(event.composed, 'event.composed');
  });
});
</script>
